requirements.txt
Django==1.7
django-compressor==1.4
djangorestframework==2.4.2
psycopg2==2.5.4
